feat: extract shared runner infrastructure, refactor all runners, extract CLI subcommands
## agents/runner-shared/ — new shared layer
- drivers/types.ts — Driver interface, DriverConfig, DriverError
- drivers/index.ts — createDriver() factory for 5 runtimes
- drivers/pi.ts — Pi CLI spawn with timeout + JSON unwrap
- drivers/anthropic.ts — Claude Agent SDK with AbortController
- drivers/cursor.ts — Cursor SDK with centralised parseCursorEvent()
- drivers/opencode.ts — OpenCode SDK with session lifecycle
- drivers/codex.ts — OpenAI Agents SDK with atomic OAuth refresh
- parse.ts — shared argv parser (replaces 10 hand-rolled copies)
- parse.test.ts — 6 unit tests
- prompt.ts — formatPrompt() / formatSinglePrompt() helpers
- seo-prompt.ts — buildSeoPrompt() for SEO review workflows
- pool.ts — mapLimit() for bounded concurrency (max 3)
- output.ts — findRepoRoot() + writeReport() with safe paths

## Refactored runners (all <45 lines)
- agents/research-runner/{pi,anthropic,cursor,opencode,codex}/runner.ts
  - use createDriver(), parseArgs, formatPrompt, mapLimit(…, 3)
- agents/seo-review-runner/{pi,anthropic,cursor,opencode,codex}/runner.ts
  - use createDriver(), parseArgs, buildSeoPrompt, writeReport
- deleted: agents/seo-review-runner/shared/prompt.ts

## CLI extraction
- packages/cli/src/cli/commands/{types,init,install,clean,reinstall,ralph,workflow}.ts
- packages/cli/src/cli/run.ts — reduced from 768 to ~150 lines (dispatcher only)
- deleted: packages/cli/src/execution/{cli-output.ts,output.ts} (orphan abstractions)

## Fixes
- agents/research-runner/shared/output.ts — VAULT_PATH now required (no personal fallback)
- workflowCommand — better error message when npx spawn fails
