+++ agents/seo-review-runner/opencode/runner.ts
@@ -7,77 +7,36 @@
  *   npx tsx runner.ts --agent technical-seo "https://example.com"
  */
 
-import { createOpencode, type OpencodeClient } from "@opencode-ai/sdk";
-import { buildPrompt, parseArgs, writeReport } from "../shared/prompt.ts";
-
-async function autoApprovePermissions(client: OpencodeClient): Promise<void> {
-  try {
-    const stream = await client.global.event();
-    stream.on("message", async (event: unknown) => {
-      const data = (
-        event as { data?: { type?: string; sessionId?: string; permissionId?: string } }
-      ).data;
-      if (data?.type === "permission" && data.sessionId && data.permissionId) {
-        try {
-          await client.postSessionIdPermissionsPermissionId({
-            path: { id: data.sessionId, permissionId: data.permissionId },
-            body: { response: "always" },
-          });
-        } catch {
-          // Non-fatal
-        }
-      }
-    });
-  } catch {
-    console.error("Warning: could not subscribe to event stream for auto-approve");
-  }
-}
+import { createDriver } from "../../runner-shared/drivers/index.js";
+import { parseArgs } from "../../runner-shared/parse.js";
+import { buildSeoPrompt } from "../../runner-shared/seo-prompt.js";
+import { writeReport } from "../../runner-shared/output.js";
 
 async function main(): Promise<void> {
-  const { url, agent } = parseArgs();
+  const { positionals, flags } = parseArgs(process.argv.slice(2), ["--agent"]);
+
+  const url = positionals.join(" ").trim();
   if (!url) {
     console.error('Usage: npx tsx runner.ts [--agent technical-seo] "https://example.com"');
     process.exit(1);
   }
 
-  const prompt = buildPrompt(url, agent);
-  console.error(`Starting OpenCode server…`);
-  const { client, server } = await createOpencode();
+  const agent = (flags["--agent"] as string | undefined) ??
+    process.env.AI_ENG_AGENT?.trim() ||
+    undefined;
 
-  await autoApprovePermissions(client);
+  const prompt = buildSeoPrompt(url, agent);
+  console.error(`Running SEO review via OpenCode driver for ${url}…`);
 
-  console.error(`Running SEO review via OpenCode SDK for ${url}…`);
-
-  const sessionResp = await client.session.create({
-    body: { title: `seo-review: ${url}` },
-  });
-
-  if (!sessionResp.data) {
-    server.close();
-    throw new Error(`Failed to create session: ${JSON.stringify(sessionResp.error)}`);
+  const driver = await createDriver("opencode");
+  try {
+    const report = await driver.runPrompt(prompt);
+    const reportPath = writeReport(url, report, "opencode", "seo-review");
+    console.error(`SEO review written to: ${reportPath}`);
+    console.log(report);
+  } finally {
+    await driver.close?.();
   }
-
-  const sessionId = sessionResp.data.id;
-
-  const promptResp = await client.session.prompt({
-    path: { id: sessionId },
-    body: { parts: [{ type: "text", text: prompt }] },
-  } as Parameters<typeof client.session.prompt>[0]);
-
-  server.close();
-
-  if (!promptResp.data) {
-    throw new Error(`Prompt failed: ${JSON.stringify(promptResp.error)}`);
-  }
-
-  const textPart = (promptResp.data as any).parts?.find(
-    (p: { type: string }) => p.type === "text",
-  );
-  const report: string = (textPart as { text?: string } | undefined)?.text ?? "(no text response)";
-
-  const reportPath = writeReport(url, report.trim() || "(no report)", "opencode");
-  console.error(`SEO review written to: ${reportPath}`);
-  console.log(report);
 }
 
 main().catch((err) => {
