+++ packages/cli/src/cli/commands/types.ts
@@ -1,0 +1,10 @@
+/**
+ * Subcommand contract for ai-eng CLI.
+ */
+
+export interface Subcommand {
+  name: string;
+  aliases?: string[];
+  helpText: string;
+  run(args: string[]): Promise<void>;
+}
